Duplicating Cabins

+++ 17-the-wild-oasis/src/services/apiCabins.js
@@ -21,9 +21,10 @@ export async function deleteCabin(id) {
 }
 
 export async function createEditCabin(newCabin, id) {
+    console.log(newCabin);
     const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
-    const imageName = `${Math.random()}-${newCabin.image.name}`.replace(
+    const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
         "/",
         ""
     );
@@ -31,15 +32,14 @@ export async function createEditCabin(newCabin, id) {
         ? newCabin.image
         : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-    // https://esitrbebvfuklmywupyu.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg
-
     // 1. Create/edit cabin
     let query = supabase.from("cabins");
 
-    //  A) create
+    // A) CREATE
     if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
 
-    // B) edit
+    // B) EDIT
+
     if (id)
         query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
 
@@ -51,21 +51,20 @@ export async function createEditCabin(newCabin, id) {
     }
 
     // 2. Upload image
+    if (hasImagePath) return data;
+
     const { error: storageError } = await supabase.storage
         .from("cabin-images")
         .upload(imageName, newCabin.image);
 
-    // 3. Delete she cabin if there was an error uploading image
+    // 3. Delete the cabin IF there was an error uplaoding image
     if (storageError) {
         await supabase.from("cabins").delete().eq("id", data.id);
-        console.log(storageError);
+        console.error(storageError);
         throw new Error(
-            "Cabin image could not be uploaded and the cabin was not created."
+            "Cabin image could not be uploaded and the cabin was not created"
         );
     }
 
-    if (storageError) {
-        console.error(storageError);
-        throw new Error("Cabins could not be created.");
-    }
+    return data;
 }
